2022 - day 7

+++ 2022/day07/test.h
@@ -3,6 +3,6 @@
 #include <string>
 
 namespace aoc2022_day07 {
-    int part_1(std::string_view path);
-    int part_2(std::string_view path);
+    long long part_1(std::string_view path);
+    long long part_2(std::string_view path);
 }
